+++ index.html
@@ -7,8 +7,6 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
-  <script src="/js/functions.js"></script>
-  <script src="/js/main.js"></script>
 </head>
 
 <body>
@@ -233,5 +231,8 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
+
+  <script src="/js/functions.js"></script>
+  <script src="/js/main.js" type="module"></script>
 </body>
 </html>
