+++ js/modules/constant.js
@@ -1,0 +1,32 @@
+export const PHOTO_COUNT = 25;
+export const AVATAR_COUNT = 6;
+export const NAMES = ['Ваня', 'Петя', 'Саша', 'Маша', 'Таня', 'Аня', 'Игорь', 'Илья', 'Вова', 'Кристина', 'Катя', 'Дима', 'Ким', 'Аким', 'Степан', 'Кирилл', 'Антон', 'Гоша', 'Виталя', 'Данил', 'Денис', 'Полина', 'Арина', 'Даша'];
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  'Я залип на этой фотке и не могу оторваться. Совсем не знаю что мне делать.',
+  'Горизонт завален.',
+  'Нормас.',
+  'Шоб я так жил!',
+  'Фокус размыт. Или это просто кто-то заляпал объектив?',
+  'Непонятен один момент: как так-то?!',
+  'Всё отлично!',
+  'Разве это композиция?! Что это за композиция?!',
+  'Кадрирование просто никакое.'
+];
+export const DESCRIPTIONS = [
+  'Классная фотка',
+  'Красивый вид',
+  'Это фото запомнится мне надолго...',
+  '',
+  'Вау!',
+  'Круто провел время',
+  'Было классно',
+  'Цените каждое мгновенье.',
+  'Тестим новую камеру!',
+  'Верьте в себя. Главное хотеть и мечтать...'
+];
